test: migrate about page tests to app router
