+++ __tests__/app/about/mission/index.test.tsx
@@ -1,16 +1,17 @@
 import React from 'react';
+/* eslint-disable react/display-name */
 import { render, screen } from '@testing-library/react';
-import MissionPage from '../../../../pages/about/mission/index';
+import MissionPage from '@/app/about/mission/page';
 
 // Mock the Header component since it's dynamically imported
-jest.mock('../../../../components/header/Header', () => {
+jest.mock('@/components/header/Header', () => {
   return function MockHeader() {
     return <div data-testid="header">Header</div>;
   };
 });
 
 // Mock HeaderPlaceholder
-jest.mock('../../../../components/header/HeaderPlaceholder', () => {
+jest.mock('@/components/header/HeaderPlaceholder', () => {
   return function MockHeaderPlaceholder() {
     return <div data-testid="header-placeholder">Header Placeholder</div>;
   };
